First working cosmo sim. Grav errors are below a few percent with 0.005 of TREE_OPEN_PARAM_REL, even in the linear regime. Added automatic adding of sourcefiles in Makefile. Added GADGET_COMOVING_VEL_UNIT to read the conventional sqrt(a) gadget snapshots. However, we use the correct a^2 dependency in the snapshots, time to cut that legacy behaviour.

+++ src/periodic.c
@@ -2,7 +2,7 @@
 
 #ifdef PERIODIC
 
-double Boxsize = 0, Boxhalf = 0;
+Float Boxsize = 0, Boxhalf = 0;
 
 void Init_Periodic()
 {
@@ -16,9 +16,9 @@ void Init_Periodic()
 
 void Periodic_Constrain_Particles_To_Box()
 {
-	const Float boxsize[3] = {Sim.Boxsize[0], Sim.Boxsize[1], Sim.Boxsize[2]};
+	const Float boxsize[3] = { Sim.Boxsize[0], Sim.Boxsize[1], Sim.Boxsize[2] };
 
-	#pragma omp for
+	#pragma omp for 
 	for (int ipart = 0; ipart < Task.Npart_Total; ipart++) {
 
 		while (P[ipart].Pos[0] < 0)
@@ -45,10 +45,10 @@ void Periodic_Constrain_Particles_To_Box()
 
 /*
  * Do the periodic mapping on a 3D distance array. We are relying on link time 
- * optimization of the compiler to do the inlining for us. 
+ * optimization of the compiler to do the inlining & optimization for us. 
  */
 
-void Periodic_Nearest(double dr[3])
+void Periodic_Nearest(Float dr[3])
 {
 	if (dr[0] > Boxhalf)
 		dr[0] -= Boxsize;
